feat(energy): persist trip battery state and observed checkpoints

+++ src/main/java/com/navio/tripplanningservice/dto/TripEvOptimizationPreviewResponse.java
@@ -8,9 +8,9 @@ public record TripEvOptimizationPreviewResponse(
         String blockId,
         boolean feasible,
         List<Operation> operations,
-        int finalSocPct,
+        Double finalSocPct,
         long totalDrivingSeconds,
-        int totalChargingMinutes,
+        double totalChargingMinutes,
         String message,
         List<String> warnings
 ) {
@@ -25,9 +25,9 @@ public record TripEvOptimizationPreviewResponse(
             String beforeItemId,
             int sequence,
             Charger charger,
-            int estimatedChargeMinutes,
-            int arrivalSocPct,
-            int departureSocPct,
+            double estimatedChargeMinutes,
+            double arrivalSocPct,
+            double departureSocPct,
             double detourKm,
             String reason
     ) {
